@@ -6,70 +6,86 @@
 # Changes may cause incorrect behavior and will be lost if the code is regenerated.
 # --------------------------------------------------------------------------
 
-from typing import Any, Optional, TYPE_CHECKING
+from copy import deepcopy
+from typing import Any, Awaitable, TYPE_CHECKING
 
+from azure.core.rest import AsyncHttpResponse, HttpRequest
 from azure.mgmt.core import AsyncARMPipelineClient
-from msrest import Deserializer, Serializer
+
+from .. import models
+from .._serialization import Deserializer, Serializer
+from ._configuration import AlertsManagementClientConfiguration
+from .operations import AlertProcessingRulesOperations, AlertsOperations, Operations, SmartGroupsOperations
 
 if TYPE_CHECKING:
     # pylint: disable=unused-import,ungrouped-imports
     from azure.core.credentials_async import AsyncTokenCredential
 
-from ._configuration import AlertsManagementClientConfiguration
-from .operations import ActionRulesOperations
-from .operations import Operations
-from .operations import AlertsOperations
-from .operations import SmartGroupsOperations
-from .operations import SmartDetectorAlertRulesOperations
-from .. import models
 
-
-class AlertsManagementClient(object):
+class AlertsManagementClient:  # pylint: disable=client-accepts-api-version-keyword
     """AlertsManagement Client.
 
-    :ivar action_rules: ActionRulesOperations operations
-    :vartype action_rules: azure.mgmt.alertsmanagement.aio.operations.ActionRulesOperations
+    :ivar alert_processing_rules: AlertProcessingRulesOperations operations
+    :vartype alert_processing_rules:
+     azure.mgmt.alertsmanagement.aio.operations.AlertProcessingRulesOperations
     :ivar operations: Operations operations
     :vartype operations: azure.mgmt.alertsmanagement.aio.operations.Operations
     :ivar alerts: AlertsOperations operations
     :vartype alerts: azure.mgmt.alertsmanagement.aio.operations.AlertsOperations
     :ivar smart_groups: SmartGroupsOperations operations
     :vartype smart_groups: azure.mgmt.alertsmanagement.aio.operations.SmartGroupsOperations
-    :ivar smart_detector_alert_rules: SmartDetectorAlertRulesOperations operations
-    :vartype smart_detector_alert_rules: azure.mgmt.alertsmanagement.aio.operations.SmartDetectorAlertRulesOperations
-    :param credential: Credential needed for the client to connect to Azure.
+    :param credential: Credential needed for the client to connect to Azure. Required.
     :type credential: ~azure.core.credentials_async.AsyncTokenCredential
-    :param subscription_id: The ID of the target subscription.
+    :param subscription_id: The ID of the target subscription. Required.
     :type subscription_id: str
-    :param str base_url: Service URL
+    :param base_url: Service URL. Default value is "https://management.azure.com".
+    :type base_url: str
     """
 
     def __init__(
         self,
         credential: "AsyncTokenCredential",
         subscription_id: str,
-        base_url: Optional[str] = None,
+        base_url: str = "https://management.azure.com",
         **kwargs: Any
     ) -> None:
-        if not base_url:
-            base_url = 'https://management.azure.com'
-        self._config = AlertsManagementClientConfiguration(credential, subscription_id, **kwargs)
+        self._config = AlertsManagementClientConfiguration(
+            credential=credential, subscription_id=subscription_id, **kwargs
+        )
         self._client = AsyncARMPipelineClient(base_url=base_url, config=self._config, **kwargs)
 
         client_models = {k: v for k, v in models.__dict__.items() if isinstance(v, type)}
         self._serialize = Serializer(client_models)
         self._deserialize = Deserializer(client_models)
+        self._serialize.client_side_validation = False
+        self.alert_processing_rules = AlertProcessingRulesOperations(
+            self._client, self._config, self._serialize, self._deserialize
+        )
+        self.operations = Operations(self._client, self._config, self._serialize, self._deserialize)
+        self.alerts = AlertsOperations(self._client, self._config, self._serialize, self._deserialize)
+        self.smart_groups = SmartGroupsOperations(self._client, self._config, self._serialize, self._deserialize)
 
-        self.action_rules = ActionRulesOperations(
-            self._client, self._config, self._serialize, self._deserialize)
-        self.operations = Operations(
-            self._client, self._config, self._serialize, self._deserialize)
-        self.alerts = AlertsOperations(
-            self._client, self._config, self._serialize, self._deserialize)
-        self.smart_groups = SmartGroupsOperations(
-            self._client, self._config, self._serialize, self._deserialize)
-        self.smart_detector_alert_rules = SmartDetectorAlertRulesOperations(
-            self._client, self._config, self._serialize, self._deserialize)
+    def _send_request(self, request: HttpRequest, **kwargs: Any) -> Awaitable[AsyncHttpResponse]:
+        """Runs the network request through the client's chained policies.
+
+        >>> from azure.core.rest import HttpRequest
+        >>> request = HttpRequest("GET", "https://www.example.org/")
+        <HttpRequest [GET], url: 'https://www.example.org/'>
+        >>> response = await client._send_request(request)
+        <AsyncHttpResponse: 200 OK>
+
+        For more information on this code flow, see https://aka.ms/azsdk/dpcodegen/python/send_request
+
+        :param request: The network request you want to make. Required.
+        :type request: ~azure.core.rest.HttpRequest
+        :keyword bool stream: Whether the response payload will be streamed. Defaults to False.
+        :return: The response of your network call. Does not do error handling on your response.
+        :rtype: ~azure.core.rest.AsyncHttpResponse
+        """
+
+        request_copy = deepcopy(request)
+        request_copy.url = self._client.format_url(request_copy.url)
+        return self._client.send_request(request_copy, **kwargs)
 
     async def close(self) -> None:
         await self._client.close()
