@@ -10,10 +10,15 @@ from ._alerts_management_client import AlertsManagementClient
 from ._version import VERSION
 
 __version__ = VERSION
-__all__ = ['AlertsManagementClient']
 
 try:
-    from ._patch import patch_sdk  # type: ignore
-    patch_sdk()
+    from ._patch import __all__ as _patch_all
+    from ._patch import *  # type: ignore # pylint: disable=unused-wildcard-import
 except ImportError:
-    pass
+    _patch_all = []
+from ._patch import patch_sdk as _patch_sdk
+
+__all__ = ["AlertsManagementClient"]
+__all__.extend([p for p in _patch_all if p not in __all__])
+
+_patch_sdk()
