@@ -7,4 +7,15 @@
 # --------------------------------------------------------------------------
 
 from ._alerts_management_client import AlertsManagementClient
-__all__ = ['AlertsManagementClient']
+
+try:
+    from ._patch import __all__ as _patch_all
+    from ._patch import *  # type: ignore # pylint: disable=unused-wildcard-import
+except ImportError:
+    _patch_all = []
+from ._patch import patch_sdk as _patch_sdk
+
+__all__ = ["AlertsManagementClient"]
+__all__.extend([p for p in _patch_all if p not in __all__])
+
+_patch_sdk()
