@@ -6,16 +6,20 @@
 # Changes may cause incorrect behavior and will be lost if the code is regenerated.
 # --------------------------------------------------------------------------
 
-from ._action_rules_operations import ActionRulesOperations
+from ._alert_processing_rules_operations import AlertProcessingRulesOperations
 from ._operations import Operations
 from ._alerts_operations import AlertsOperations
 from ._smart_groups_operations import SmartGroupsOperations
-from ._smart_detector_alert_rules_operations import SmartDetectorAlertRulesOperations
+
+from ._patch import __all__ as _patch_all
+from ._patch import *  # type: ignore # pylint: disable=unused-wildcard-import
+from ._patch import patch_sdk as _patch_sdk
 
 __all__ = [
-    'ActionRulesOperations',
-    'Operations',
-    'AlertsOperations',
-    'SmartGroupsOperations',
-    'SmartDetectorAlertRulesOperations',
+    "AlertProcessingRulesOperations",
+    "Operations",
+    "AlertsOperations",
+    "SmartGroupsOperations",
 ]
+__all__.extend([p for p in _patch_all if p not in __all__])
+_patch_sdk()
