@@ -9,15 +9,22 @@ from typing import Any, AsyncIterable, Callable, Dict, Generic, Optional, TypeVa
 import warnings
 
 from azure.core.async_paging import AsyncItemPaged, AsyncList
-from azure.core.exceptions import ClientAuthenticationError, HttpResponseError, ResourceExistsError, ResourceNotFoundError, map_error
+from azure.core.exceptions import (
+    ClientAuthenticationError,
+    HttpResponseError,
+    ResourceExistsError,
+    ResourceNotFoundError,
+    map_error,
+)
 from azure.core.pipeline import PipelineResponse
 from azure.core.pipeline.transport import AsyncHttpResponse, HttpRequest
 from azure.mgmt.core.exceptions import ARMErrorFormat
 
 from ... import models as _models
 
-T = TypeVar('T')
+T = TypeVar("T")
 ClsType = Optional[Callable[[PipelineResponse[HttpRequest, AsyncHttpResponse], T, Dict[str, Any]], Any]]
+
 
 class ActionRulesOperations:
     """ActionRulesOperations async operations.
@@ -88,49 +95,57 @@ class ActionRulesOperations:
         :rtype: ~azure.core.async_paging.AsyncItemPaged[~azure.mgmt.alertsmanagement.models.ActionRulesList]
         :raises: ~azure.core.exceptions.HttpResponseError
         """
-        cls = kwargs.pop('cls', None)  # type: ClsType["_models.ActionRulesList"]
-        error_map = {
-            401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
-        }
-        error_map.update(kwargs.pop('error_map', {}))
+        cls = kwargs.pop("cls", None)  # type: ClsType["_models.ActionRulesList"]
+        error_map = {401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError}
+        error_map.update(kwargs.pop("error_map", {}))
         api_version = "2019-05-05-preview"
         accept = "application/json"
 
         def prepare_request(next_link=None):
             # Construct headers
             header_parameters = {}  # type: Dict[str, Any]
-            header_parameters['Accept'] = self._serialize.header("accept", accept, 'str')
+            header_parameters["Accept"] = self._serialize.header("accept", accept, "str")
 
             if not next_link:
                 # Construct URL
-                url = self.list_by_subscription.metadata['url']  # type: ignore
+                url = self.list_by_subscription.metadata["url"]  # type: ignore
                 path_format_arguments = {
-                    'subscriptionId': self._serialize.url("self._config.subscription_id", self._config.subscription_id, 'str', min_length=1),
+                    "subscriptionId": self._serialize.url(
+                        "self._config.subscription_id", self._config.subscription_id, "str", min_length=1
+                    ),
                 }
                 url = self._client.format_url(url, **path_format_arguments)
                 # Construct parameters
                 query_parameters = {}  # type: Dict[str, Any]
                 if target_resource_group is not None:
-                    query_parameters['targetResourceGroup'] = self._serialize.query("target_resource_group", target_resource_group, 'str')
+                    query_parameters["targetResourceGroup"] = self._serialize.query(
+                        "target_resource_group", target_resource_group, "str"
+                    )
                 if target_resource_type is not None:
-                    query_parameters['targetResourceType'] = self._serialize.query("target_resource_type", target_resource_type, 'str')
+                    query_parameters["targetResourceType"] = self._serialize.query(
+                        "target_resource_type", target_resource_type, "str"
+                    )
                 if target_resource is not None:
-                    query_parameters['targetResource'] = self._serialize.query("target_resource", target_resource, 'str')
+                    query_parameters["targetResource"] = self._serialize.query(
+                        "target_resource", target_resource, "str"
+                    )
                 if severity is not None:
-                    query_parameters['severity'] = self._serialize.query("severity", severity, 'str')
+                    query_parameters["severity"] = self._serialize.query("severity", severity, "str")
                 if monitor_service is not None:
-                    query_parameters['monitorService'] = self._serialize.query("monitor_service", monitor_service, 'str')
+                    query_parameters["monitorService"] = self._serialize.query(
+                        "monitor_service", monitor_service, "str"
+                    )
                 if impacted_scope is not None:
-                    query_parameters['impactedScope'] = self._serialize.query("impacted_scope", impacted_scope, 'str')
+                    query_parameters["impactedScope"] = self._serialize.query("impacted_scope", impacted_scope, "str")
                 if description is not None:
-                    query_parameters['description'] = self._serialize.query("description", description, 'str')
+                    query_parameters["description"] = self._serialize.query("description", description, "str")
                 if alert_rule_id is not None:
-                    query_parameters['alertRuleId'] = self._serialize.query("alert_rule_id", alert_rule_id, 'str')
+                    query_parameters["alertRuleId"] = self._serialize.query("alert_rule_id", alert_rule_id, "str")
                 if action_group is not None:
-                    query_parameters['actionGroup'] = self._serialize.query("action_group", action_group, 'str')
+                    query_parameters["actionGroup"] = self._serialize.query("action_group", action_group, "str")
                 if name is not None:
-                    query_parameters['name'] = self._serialize.query("name", name, 'str')
-                query_parameters['api-version'] = self._serialize.query("api_version", api_version, 'str')
+                    query_parameters["name"] = self._serialize.query("name", name, "str")
+                query_parameters["api-version"] = self._serialize.query("api_version", api_version, "str")
 
                 request = self._client.get(url, query_parameters, header_parameters)
             else:
@@ -140,7 +155,7 @@ class ActionRulesOperations:
             return request
 
         async def extract_data(pipeline_response):
-            deserialized = self._deserialize('ActionRulesList', pipeline_response)
+            deserialized = self._deserialize("ActionRulesList", pipeline_response)
             list_of_elem = deserialized.value
             if cls:
                 list_of_elem = cls(list_of_elem)
@@ -159,10 +174,9 @@ class ActionRulesOperations:
 
             return pipeline_response
 
-        return AsyncItemPaged(
-            get_next, extract_data
-        )
-    list_by_subscription.metadata = {'url': '/subscriptions/{subscriptionId}/providers/Microsoft.AlertsManagement/actionRules'}  # type: ignore
+        return AsyncItemPaged(get_next, extract_data)
+
+    list_by_subscription.metadata = {"url": "/subscriptions/{subscriptionId}/providers/Microsoft.AlertsManagement/actionRules"}  # type: ignore
 
     def list_by_resource_group(
         self,
@@ -215,50 +229,58 @@ class ActionRulesOperations:
         :rtype: ~azure.core.async_paging.AsyncItemPaged[~azure.mgmt.alertsmanagement.models.ActionRulesList]
         :raises: ~azure.core.exceptions.HttpResponseError
         """
-        cls = kwargs.pop('cls', None)  # type: ClsType["_models.ActionRulesList"]
-        error_map = {
-            401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
-        }
-        error_map.update(kwargs.pop('error_map', {}))
+        cls = kwargs.pop("cls", None)  # type: ClsType["_models.ActionRulesList"]
+        error_map = {401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError}
+        error_map.update(kwargs.pop("error_map", {}))
         api_version = "2019-05-05-preview"
         accept = "application/json"
 
         def prepare_request(next_link=None):
             # Construct headers
             header_parameters = {}  # type: Dict[str, Any]
-            header_parameters['Accept'] = self._serialize.header("accept", accept, 'str')
+            header_parameters["Accept"] = self._serialize.header("accept", accept, "str")
 
             if not next_link:
                 # Construct URL
-                url = self.list_by_resource_group.metadata['url']  # type: ignore
+                url = self.list_by_resource_group.metadata["url"]  # type: ignore
                 path_format_arguments = {
-                    'subscriptionId': self._serialize.url("self._config.subscription_id", self._config.subscription_id, 'str', min_length=1),
-                    'resourceGroupName': self._serialize.url("resource_group_name", resource_group_name, 'str'),
+                    "subscriptionId": self._serialize.url(
+                        "self._config.subscription_id", self._config.subscription_id, "str", min_length=1
+                    ),
+                    "resourceGroupName": self._serialize.url("resource_group_name", resource_group_name, "str"),
                 }
                 url = self._client.format_url(url, **path_format_arguments)
                 # Construct parameters
                 query_parameters = {}  # type: Dict[str, Any]
                 if target_resource_group is not None:
-                    query_parameters['targetResourceGroup'] = self._serialize.query("target_resource_group", target_resource_group, 'str')
+                    query_parameters["targetResourceGroup"] = self._serialize.query(
+                        "target_resource_group", target_resource_group, "str"
+                    )
                 if target_resource_type is not None:
-                    query_parameters['targetResourceType'] = self._serialize.query("target_resource_type", target_resource_type, 'str')
+                    query_parameters["targetResourceType"] = self._serialize.query(
+                        "target_resource_type", target_resource_type, "str"
+                    )
                 if target_resource is not None:
-                    query_parameters['targetResource'] = self._serialize.query("target_resource", target_resource, 'str')
+                    query_parameters["targetResource"] = self._serialize.query(
+                        "target_resource", target_resource, "str"
+                    )
                 if severity is not None:
-                    query_parameters['severity'] = self._serialize.query("severity", severity, 'str')
+                    query_parameters["severity"] = self._serialize.query("severity", severity, "str")
                 if monitor_service is not None:
-                    query_parameters['monitorService'] = self._serialize.query("monitor_service", monitor_service, 'str')
+                    query_parameters["monitorService"] = self._serialize.query(
+                        "monitor_service", monitor_service, "str"
+                    )
                 if impacted_scope is not None:
-                    query_parameters['impactedScope'] = self._serialize.query("impacted_scope", impacted_scope, 'str')
+                    query_parameters["impactedScope"] = self._serialize.query("impacted_scope", impacted_scope, "str")
                 if description is not None:
-                    query_parameters['description'] = self._serialize.query("description", description, 'str')
+                    query_parameters["description"] = self._serialize.query("description", description, "str")
                 if alert_rule_id is not None:
-                    query_parameters['alertRuleId'] = self._serialize.query("alert_rule_id", alert_rule_id, 'str')
+                    query_parameters["alertRuleId"] = self._serialize.query("alert_rule_id", alert_rule_id, "str")
                 if action_group is not None:
-                    query_parameters['actionGroup'] = self._serialize.query("action_group", action_group, 'str')
+                    query_parameters["actionGroup"] = self._serialize.query("action_group", action_group, "str")
                 if name is not None:
-                    query_parameters['name'] = self._serialize.query("name", name, 'str')
-                query_parameters['api-version'] = self._serialize.query("api_version", api_version, 'str')
+                    query_parameters["name"] = self._serialize.query("name", name, "str")
+                query_parameters["api-version"] = self._serialize.query("api_version", api_version, "str")
 
                 request = self._client.get(url, query_parameters, header_parameters)
             else:
@@ -268,7 +290,7 @@ class ActionRulesOperations:
             return request
 
         async def extract_data(pipeline_response):
-            deserialized = self._deserialize('ActionRulesList', pipeline_response)
+            deserialized = self._deserialize("ActionRulesList", pipeline_response)
             list_of_elem = deserialized.value
             if cls:
                 list_of_elem = cls(list_of_elem)
@@ -287,17 +309,11 @@ class ActionRulesOperations:
 
             return pipeline_response
 
-        return AsyncItemPaged(
-            get_next, extract_data
-        )
-    list_by_resource_group.metadata = {'url': '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.AlertsManagement/actionRules'}  # type: ignore
+        return AsyncItemPaged(get_next, extract_data)
 
-    async def get_by_name(
-        self,
-        resource_group_name: str,
-        action_rule_name: str,
-        **kwargs
-    ) -> "_models.ActionRule":
+    list_by_resource_group.metadata = {"url": "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.AlertsManagement/actionRules"}  # type: ignore
+
+    async def get_by_name(self, resource_group_name: str, action_rule_name: str, **kwargs) -> "_models.ActionRule":
         """Get action rule by name.
 
         Get a specific action rule.
@@ -311,30 +327,30 @@ class ActionRulesOperations:
         :rtype: ~azure.mgmt.alertsmanagement.models.ActionRule
         :raises: ~azure.core.exceptions.HttpResponseError
         """
-        cls = kwargs.pop('cls', None)  # type: ClsType["_models.ActionRule"]
-        error_map = {
-            401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
-        }
-        error_map.update(kwargs.pop('error_map', {}))
+        cls = kwargs.pop("cls", None)  # type: ClsType["_models.ActionRule"]
+        error_map = {401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError}
+        error_map.update(kwargs.pop("error_map", {}))
         api_version = "2019-05-05-preview"
         accept = "application/json"
 
         # Construct URL
-        url = self.get_by_name.metadata['url']  # type: ignore
+        url = self.get_by_name.metadata["url"]  # type: ignore
         path_format_arguments = {
-            'subscriptionId': self._serialize.url("self._config.subscription_id", self._config.subscription_id, 'str', min_length=1),
-            'resourceGroupName': self._serialize.url("resource_group_name", resource_group_name, 'str'),
-            'actionRuleName': self._serialize.url("action_rule_name", action_rule_name, 'str'),
+            "subscriptionId": self._serialize.url(
+                "self._config.subscription_id", self._config.subscription_id, "str", min_length=1
+            ),
+            "resourceGroupName": self._serialize.url("resource_group_name", resource_group_name, "str"),
+            "actionRuleName": self._serialize.url("action_rule_name", action_rule_name, "str"),
         }
         url = self._client.format_url(url, **path_format_arguments)
 
         # Construct parameters
         query_parameters = {}  # type: Dict[str, Any]
-        query_parameters['api-version'] = self._serialize.query("api_version", api_version, 'str')
+        query_parameters["api-version"] = self._serialize.query("api_version", api_version, "str")
 
         # Construct headers
         header_parameters = {}  # type: Dict[str, Any]
-        header_parameters['Accept'] = self._serialize.header("accept", accept, 'str')
+        header_parameters["Accept"] = self._serialize.header("accept", accept, "str")
 
         request = self._client.get(url, query_parameters, header_parameters)
         pipeline_response = await self._client._pipeline.run(request, stream=False, **kwargs)
@@ -346,21 +362,18 @@ class ActionRulesOperations:
             raise HttpResponseError(response=response, model=error, error_format=ARMErrorFormat)
 
         response_headers = {}
-        response_headers['x-ms-request-id']=self._deserialize('str', response.headers.get('x-ms-request-id'))
-        deserialized = self._deserialize('ActionRule', pipeline_response)
+        response_headers["x-ms-request-id"] = self._deserialize("str", response.headers.get("x-ms-request-id"))
+        deserialized = self._deserialize("ActionRule", pipeline_response)
 
         if cls:
             return cls(pipeline_response, deserialized, response_headers)
 
         return deserialized
-    get_by_name.metadata = {'url': '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.AlertsManagement/actionRules/{actionRuleName}'}  # type: ignore
+
+    get_by_name.metadata = {"url": "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.AlertsManagement/actionRules/{actionRuleName}"}  # type: ignore
 
     async def create_update(
-        self,
-        resource_group_name: str,
-        action_rule_name: str,
-        action_rule: "_models.ActionRule",
-        **kwargs
+        self, resource_group_name: str, action_rule_name: str, action_rule: "_models.ActionRule", **kwargs
     ) -> "_models.ActionRule":
         """Create/update an action rule.
 
@@ -377,36 +390,36 @@ class ActionRulesOperations:
         :rtype: ~azure.mgmt.alertsmanagement.models.ActionRule
         :raises: ~azure.core.exceptions.HttpResponseError
         """
-        cls = kwargs.pop('cls', None)  # type: ClsType["_models.ActionRule"]
-        error_map = {
-            401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
-        }
-        error_map.update(kwargs.pop('error_map', {}))
+        cls = kwargs.pop("cls", None)  # type: ClsType["_models.ActionRule"]
+        error_map = {401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError}
+        error_map.update(kwargs.pop("error_map", {}))
         api_version = "2019-05-05-preview"
         content_type = kwargs.pop("content_type", "application/json")
         accept = "application/json"
 
         # Construct URL
-        url = self.create_update.metadata['url']  # type: ignore
+        url = self.create_update.metadata["url"]  # type: ignore
         path_format_arguments = {
-            'subscriptionId': self._serialize.url("self._config.subscription_id", self._config.subscription_id, 'str', min_length=1),
-            'resourceGroupName': self._serialize.url("resource_group_name", resource_group_name, 'str'),
-            'actionRuleName': self._serialize.url("action_rule_name", action_rule_name, 'str'),
+            "subscriptionId": self._serialize.url(
+                "self._config.subscription_id", self._config.subscription_id, "str", min_length=1
+            ),
+            "resourceGroupName": self._serialize.url("resource_group_name", resource_group_name, "str"),
+            "actionRuleName": self._serialize.url("action_rule_name", action_rule_name, "str"),
         }
         url = self._client.format_url(url, **path_format_arguments)
 
         # Construct parameters
         query_parameters = {}  # type: Dict[str, Any]
-        query_parameters['api-version'] = self._serialize.query("api_version", api_version, 'str')
+        query_parameters["api-version"] = self._serialize.query("api_version", api_version, "str")
 
         # Construct headers
         header_parameters = {}  # type: Dict[str, Any]
-        header_parameters['Content-Type'] = self._serialize.header("content_type", content_type, 'str')
-        header_parameters['Accept'] = self._serialize.header("accept", accept, 'str')
+        header_parameters["Content-Type"] = self._serialize.header("content_type", content_type, "str")
+        header_parameters["Accept"] = self._serialize.header("accept", accept, "str")
 
         body_content_kwargs = {}  # type: Dict[str, Any]
-        body_content = self._serialize.body(action_rule, 'ActionRule')
-        body_content_kwargs['content'] = body_content
+        body_content = self._serialize.body(action_rule, "ActionRule")
+        body_content_kwargs["content"] = body_content
         request = self._client.put(url, query_parameters, header_parameters, **body_content_kwargs)
         pipeline_response = await self._client._pipeline.run(request, stream=False, **kwargs)
         response = pipeline_response.http_response
@@ -417,21 +430,17 @@ class ActionRulesOperations:
             raise HttpResponseError(response=response, model=error, error_format=ARMErrorFormat)
 
         response_headers = {}
-        response_headers['x-ms-request-id']=self._deserialize('str', response.headers.get('x-ms-request-id'))
-        deserialized = self._deserialize('ActionRule', pipeline_response)
+        response_headers["x-ms-request-id"] = self._deserialize("str", response.headers.get("x-ms-request-id"))
+        deserialized = self._deserialize("ActionRule", pipeline_response)
 
         if cls:
             return cls(pipeline_response, deserialized, response_headers)
 
         return deserialized
-    create_update.metadata = {'url': '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.AlertsManagement/actionRules/{actionRuleName}'}  # type: ignore
 
-    async def delete(
-        self,
-        resource_group_name: str,
-        action_rule_name: str,
-        **kwargs
-    ) -> bool:
+    create_update.metadata = {"url": "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.AlertsManagement/actionRules/{actionRuleName}"}  # type: ignore
+
+    async def delete(self, resource_group_name: str, action_rule_name: str, **kwargs) -> bool:
         """Delete action rule.
 
         Deletes a given action rule.
@@ -445,30 +454,30 @@ class ActionRulesOperations:
         :rtype: bool
         :raises: ~azure.core.exceptions.HttpResponseError
         """
-        cls = kwargs.pop('cls', None)  # type: ClsType[bool]
-        error_map = {
-            401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
-        }
-        error_map.update(kwargs.pop('error_map', {}))
+        cls = kwargs.pop("cls", None)  # type: ClsType[bool]
+        error_map = {401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError}
+        error_map.update(kwargs.pop("error_map", {}))
         api_version = "2019-05-05-preview"
         accept = "application/json"
 
         # Construct URL
-        url = self.delete.metadata['url']  # type: ignore
+        url = self.delete.metadata["url"]  # type: ignore
         path_format_arguments = {
-            'subscriptionId': self._serialize.url("self._config.subscription_id", self._config.subscription_id, 'str', min_length=1),
-            'resourceGroupName': self._serialize.url("resource_group_name", resource_group_name, 'str'),
-            'actionRuleName': self._serialize.url("action_rule_name", action_rule_name, 'str'),
+            "subscriptionId": self._serialize.url(
+                "self._config.subscription_id", self._config.subscription_id, "str", min_length=1
+            ),
+            "resourceGroupName": self._serialize.url("resource_group_name", resource_group_name, "str"),
+            "actionRuleName": self._serialize.url("action_rule_name", action_rule_name, "str"),
         }
         url = self._client.format_url(url, **path_format_arguments)
 
         # Construct parameters
         query_parameters = {}  # type: Dict[str, Any]
-        query_parameters['api-version'] = self._serialize.query("api_version", api_version, 'str')
+        query_parameters["api-version"] = self._serialize.query("api_version", api_version, "str")
 
         # Construct headers
         header_parameters = {}  # type: Dict[str, Any]
-        header_parameters['Accept'] = self._serialize.header("accept", accept, 'str')
+        header_parameters["Accept"] = self._serialize.header("accept", accept, "str")
 
         request = self._client.delete(url, query_parameters, header_parameters)
         pipeline_response = await self._client._pipeline.run(request, stream=False, **kwargs)
@@ -480,21 +489,18 @@ class ActionRulesOperations:
             raise HttpResponseError(response=response, model=error, error_format=ARMErrorFormat)
 
         response_headers = {}
-        response_headers['x-ms-request-id']=self._deserialize('str', response.headers.get('x-ms-request-id'))
-        deserialized = self._deserialize('bool', pipeline_response)
+        response_headers["x-ms-request-id"] = self._deserialize("str", response.headers.get("x-ms-request-id"))
+        deserialized = self._deserialize("bool", pipeline_response)
 
         if cls:
             return cls(pipeline_response, deserialized, response_headers)
 
         return deserialized
-    delete.metadata = {'url': '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.AlertsManagement/actionRules/{actionRuleName}'}  # type: ignore
+
+    delete.metadata = {"url": "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.AlertsManagement/actionRules/{actionRuleName}"}  # type: ignore
 
     async def update(
-        self,
-        resource_group_name: str,
-        action_rule_name: str,
-        action_rule_patch: "_models.PatchObject",
-        **kwargs
+        self, resource_group_name: str, action_rule_name: str, action_rule_patch: "_models.PatchObject", **kwargs
     ) -> "_models.ActionRule":
         """Patch action rule.
 
@@ -511,36 +517,36 @@ class ActionRulesOperations:
         :rtype: ~azure.mgmt.alertsmanagement.models.ActionRule
         :raises: ~azure.core.exceptions.HttpResponseError
         """
-        cls = kwargs.pop('cls', None)  # type: ClsType["_models.ActionRule"]
-        error_map = {
-            401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
-        }
-        error_map.update(kwargs.pop('error_map', {}))
+        cls = kwargs.pop("cls", None)  # type: ClsType["_models.ActionRule"]
+        error_map = {401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError}
+        error_map.update(kwargs.pop("error_map", {}))
         api_version = "2019-05-05-preview"
         content_type = kwargs.pop("content_type", "application/json")
         accept = "application/json"
 
         # Construct URL
-        url = self.update.metadata['url']  # type: ignore
+        url = self.update.metadata["url"]  # type: ignore
         path_format_arguments = {
-            'subscriptionId': self._serialize.url("self._config.subscription_id", self._config.subscription_id, 'str', min_length=1),
-            'resourceGroupName': self._serialize.url("resource_group_name", resource_group_name, 'str'),
-            'actionRuleName': self._serialize.url("action_rule_name", action_rule_name, 'str'),
+            "subscriptionId": self._serialize.url(
+                "self._config.subscription_id", self._config.subscription_id, "str", min_length=1
+            ),
+            "resourceGroupName": self._serialize.url("resource_group_name", resource_group_name, "str"),
+            "actionRuleName": self._serialize.url("action_rule_name", action_rule_name, "str"),
         }
         url = self._client.format_url(url, **path_format_arguments)
 
         # Construct parameters
         query_parameters = {}  # type: Dict[str, Any]
-        query_parameters['api-version'] = self._serialize.query("api_version", api_version, 'str')
+        query_parameters["api-version"] = self._serialize.query("api_version", api_version, "str")
 
         # Construct headers
         header_parameters = {}  # type: Dict[str, Any]
-        header_parameters['Content-Type'] = self._serialize.header("content_type", content_type, 'str')
-        header_parameters['Accept'] = self._serialize.header("accept", accept, 'str')
+        header_parameters["Content-Type"] = self._serialize.header("content_type", content_type, "str")
+        header_parameters["Accept"] = self._serialize.header("accept", accept, "str")
 
         body_content_kwargs = {}  # type: Dict[str, Any]
-        body_content = self._serialize.body(action_rule_patch, 'PatchObject')
-        body_content_kwargs['content'] = body_content
+        body_content = self._serialize.body(action_rule_patch, "PatchObject")
+        body_content_kwargs["content"] = body_content
         request = self._client.patch(url, query_parameters, header_parameters, **body_content_kwargs)
         pipeline_response = await self._client._pipeline.run(request, stream=False, **kwargs)
         response = pipeline_response.http_response
@@ -551,11 +557,12 @@ class ActionRulesOperations:
             raise HttpResponseError(response=response, model=error, error_format=ARMErrorFormat)
 
         response_headers = {}
-        response_headers['x-ms-request-id']=self._deserialize('str', response.headers.get('x-ms-request-id'))
-        deserialized = self._deserialize('ActionRule', pipeline_response)
+        response_headers["x-ms-request-id"] = self._deserialize("str", response.headers.get("x-ms-request-id"))
+        deserialized = self._deserialize("ActionRule", pipeline_response)
 
         if cls:
             return cls(pipeline_response, deserialized, response_headers)
 
         return deserialized
-    update.metadata = {'url': '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.AlertsManagement/actionRules/{actionRuleName}'}  # type: ignore
+
+    update.metadata = {"url": "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.AlertsManagement/actionRules/{actionRuleName}"}  # type: ignore
